updated answering and evaluation scripts

+++ answer.py
@@ -38,38 +38,40 @@ for i in range(NUMBER_EXECUTIONS):
         dream_path = os.path.join("incipits", incipit)
         answer_path = os.path.join("answers", m_name+"__"+incipit.split(".")[0]+"__"+str(i)+".txt")
 
-        messages = [{"role": "user", "content": "You are dreaming. Can you complete the following dream?\n\n"+open(dream_path, "r").read()}]
+        if not os.path.exists(answer_path):
+            messages = [{"role": "user", "content": "You are dreaming. Can you complete the following dream?\n\n"+open(dream_path, "r").read()}]
 
-        headers = {
-            "Content-Type": "application/json",
-            "Authorization": f"Bearer {API_KEY}"
-        }
+            headers = {
+                "Content-Type": "application/json",
+                "Authorization": f"Bearer {API_KEY}"
+            }
 
-        payload = {
-            "model": MODEL_NAME,
-            "messages": messages,
-        }
+            payload = {
+                "model": MODEL_NAME,
+                "messages": messages,
+                "max_tokens": 4096
+            }
 
-        if payload is not None:
-            complete_url = API_URL + "chat/completions"
+            if payload is not None:
+                complete_url = API_URL + "chat/completions"
 
-            response_message = ""
-            response = None
+                response_message = ""
+                response = None
 
-            while not response_message:
-                try:
-                    response = requests.post(complete_url, headers=headers, json=payload).json()
+                while not response_message:
+                    try:
+                        response = requests.post(complete_url, headers=headers, json=payload).json()
 
-                    response_message = strip_non_unicode_characters(response["choices"][0]["message"]["content"])
+                        response_message = strip_non_unicode_characters(response["choices"][0]["message"]["content"])
 
-                    F = open(answer_path, "w")
-                    F.write(response_message)
-                    F.close()
-                except:
-                    print(response)
+                        F = open(answer_path, "w")
+                        F.write(response_message)
+                        F.close()
+                    except:
+                        print(response)
 
-                    traceback.print_exc()
+                        traceback.print_exc()
 
-                    print("sleeping %d seconds ..." % (WAITING_TIME_RETRY))
+                        print("sleeping %d seconds ..." % (WAITING_TIME_RETRY))
 
-                    time.sleep(WAITING_TIME_RETRY)
+                        time.sleep(WAITING_TIME_RETRY)
